tidy(ui): extract resolveBatchValues to own file
